Moved global optimisation into morris strategy folder
